vajra/Front-end/requirements.txt
requests==2.25.1
Django==3.2
hurry.filesize==0.9